+++ src/engine/core/objects/ai/index.ts
@@ -1,0 +1,1 @@
+export * from "@/engine/core/objects/ai/setup";
