+++ src/engine/core/objects/ai/setup/index.ts
@@ -1,0 +1,2 @@
+export * from "@/engine/core/objects/ai/setup/motivation_planner";
+export * from "@/engine/core/objects/ai/setup/state_planner";
