+++ src/engine/core/objects/ai/setup/state/couch_planner.ts
@@ -1,0 +1,9 @@
+import { StalkerStateManager } from "@/engine/core/objects/state";
+import { ActionPlanner } from "@/engine/lib/types";
+
+/**
+ *
+ * @param planner
+ * @param stateManager
+ */
+export function setupStalkerMentalStatePlanner(planner: ActionPlanner, stateManager: StalkerStateManager): void {}
