+++ src/engine/core/objects/ai/setup/state/weapon_planner.ts
@@ -1,0 +1,80 @@
+import { world_property } from "xray16";
+
+import { EStateActionId, EStateEvaluatorId } from "@/engine/core/objects/animation";
+import { StalkerStateManager } from "@/engine/core/objects/state";
+import {
+  ActionWeaponDrop,
+  ActionWeaponNone,
+  ActionWeaponStrap,
+  ActionWeaponUnstrap,
+} from "@/engine/core/objects/state/weapon";
+import { ActionPlanner } from "@/engine/lib/types";
+
+/**
+ *
+ * @param planner
+ * @param stateManager
+ */
+export function setupStalkerWeaponStatePlanner(planner: ActionPlanner, stateManager: StalkerStateManager): void {
+  // -- UNSTRAPP
+  const unstrapAction: ActionWeaponUnstrap = new ActionWeaponUnstrap(stateManager);
+
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED, false));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_LOCKED, false));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_LOCKED, false));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, false));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.MOVEMENT, true));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.BODYSTATE, true));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.MENTAL, true));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.WEAPON_UNSTRAPPED, true));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_IDLE_NOW, true));
+  unstrapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_NONE_NOW, true));
+  unstrapAction.add_effect(new world_property(EStateEvaluatorId.WEAPON, true));
+  planner.add_action(EStateActionId.WEAPON_UNSTRAPP, unstrapAction);
+
+  // -- STRAPP
+  const strapAction: ActionWeaponStrap = new ActionWeaponStrap(stateManager);
+
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED, false));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_LOCKED, false));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_LOCKED, false));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, false));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.MOVEMENT, true));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.BODYSTATE, true));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.MENTAL, true));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.WEAPON_STRAPPED, true));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_IDLE_NOW, true));
+  strapAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_NONE_NOW, true));
+  strapAction.add_effect(new world_property(EStateEvaluatorId.WEAPON, true));
+  planner.add_action(EStateActionId.WEAPON_STRAPP, strapAction);
+
+  // -- NONE
+  const weaponNoneAction: ActionWeaponNone = new ActionWeaponNone(stateManager);
+
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED, false));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_LOCKED, false));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_LOCKED, false));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, false));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.MOVEMENT, true));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.BODYSTATE, true));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.MENTAL, true));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.WEAPON_NONE, true));
+  weaponNoneAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_NONE_NOW, true));
+  weaponNoneAction.add_effect(new world_property(EStateEvaluatorId.WEAPON, true));
+  planner.add_action(EStateActionId.WEAPON_NONE, weaponNoneAction);
+
+  // -- DROP
+  const weaponDropAction: ActionWeaponDrop = new ActionWeaponDrop(stateManager);
+
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED, false));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.ANIMSTATE_LOCKED, false));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_LOCKED, false));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.LOCKED_EXTERNAL, false));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.MOVEMENT, true));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.BODYSTATE, true));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.MENTAL, true));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.WEAPON_DROP, true));
+  weaponDropAction.add_precondition(new world_property(EStateEvaluatorId.ANIMATION_NONE_NOW, true));
+  weaponDropAction.add_effect(new world_property(EStateEvaluatorId.WEAPON, true));
+  planner.add_action(EStateActionId.WEAPON_DROP, weaponDropAction);
+}
