+++ src/fixtures/engine/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "@/fixtures/engine/utils/check_action_planner";
 export * from "@/fixtures/engine/utils/check_binding";
 export * from "@/fixtures/engine/utils/check_intersection";
 export * from "@/fixtures/engine/utils/read_ltx";
